Move pure views out of the way into views.py
